+++ docs/source/changelog.rst
@@ -6,6 +6,12 @@ pyfdc 0.1.3
 
 
 * 
+  ``get_nutrients`` was dropped. Use ``get_food_info`` with ``target_field`` "nutrients."
+
+* 
+  ``get_multiple_details`` was dropped. Use ``get_food_info`` instead.
+
+* 
   Classes FoodSearch and FoodDetails have been dropped. Use ``FoodDataCentral`` instead.
 
 *